test: end-to-end write_parquet trailing-slash test; move unit test
Per review:
- Replace the writer-instance test with a full df.write_parquet roundtrip in
  tests/io/test_parquet.py — that's the user-visible bug surface.
- Move the _resolve_paths_and_filesystem unit test to tests/io/test_filesystem.py
  where filesystem-level coverage belongs.
- Tighten the rationale comment to explain *why* we strip.
