+++ daft/filesystem.py
@@ -217,7 +217,7 @@ def _infer_filesystem(
         io_config: A Daft IOConfig that should be best-effort applied onto the returned
             FileSystem
     """
-    # Strip trailing slashes (PyArrow's normalize_path doesn't).
+    # Strip trailing slashes so downstream "{dir}/{file}" joins don't double them.
     path = path.rstrip("/") or path
     protocol = get_protocol_from_path(path)
     translated_kwargs: dict[str, Any]
